Refactor the optimization pass API
