+++ onnx/optimizer/passes/fuse_transpose_into_gemm.h
@@ -1,0 +1,39 @@
+#pragma once
+
+#include "onnx/optimizer/passes/optimize_pass.h"
+
+namespace onnx { namespace optimization {
+
+struct FuseTransposeIntoGemm : public OptimizePass {
+  explicit FuseTransposeIntoGemm()
+    : OptimizePass("fuse_transpose_into_gemm", API_TYPE::ir) {
+  }
+
+  void fuse_transpose_into_gemm(std::shared_ptr<Graph>& graph) {
+    static const std::vector<int64_t> simple_trans_perm({1,0});
+
+    for (auto it = graph->begin(); it != graph->end(); ++it) {
+      auto* n = *it;
+
+      if (n->kind() == kGemm) {
+        for (size_t i : {0,1}) {
+          auto inp = n->inputs()[i];
+          auto trans = i == 0 ? ktransA : ktransB;
+          if (inp->node()->kind() == kTranspose && inp->node()->is(kperm) == simple_trans_perm) {
+            n->replaceInput(i, inp->node()->input());
+            n->i_(trans, n->hasAttribute(trans) ? !n->i(trans) : 1);
+            if (inp->uses().size() == 0) {
+              inp->node()->destroy();
+            }
+          }
+        }
+      }
+    }
+  }
+
+  virtual void optimize(std::shared_ptr<Graph>& graph) {
+    fuse_transpose_into_gemm(graph);
+  }
+};
+
+}} // namespace onnx::optimization
